# perf(integrals): use xreplace instead of subs in heurisch