@oscarbenjamin 2021-04-20
The OP example is from #3890 and the other example there `integrate(cos(x)**2 * sin(2*x), x)` is also made 4x faster with this.